spawn_app() + test created
