+++ src/main.rs
@@ -4,7 +4,6 @@ use zero2prod::run;
 
 #[tokio::main]
 async fn main() -> std::io::Result<()> {
-    let listener = TcpListener::bind("127.0.0.1:8080")
-        .expect("Failed to bind random port");
+    let listener = TcpListener::bind("127.0.0.1:8080").expect("Failed to bind random port");
     run(listener)?.await
 }
